requirements.txt
pynest-api==0.2.0
aiomysql==0.2.0
pyyaml
requests
pymodbus==2.5.3
paho-mqtt
asyncio-mqtt
pymysql
opcua
flask
fastapi
SQLAlchemy
sqlalchemy
eventlet
flask_cors
cryptography
bcrypt
flask-bcrypt
psutil
redis
pandas
uvicorn
jwt
python-jose
pydantic-settings
passlib
pydantic
pydantic[email]
python-multipart
mysql-connector-python
mybatis-mapper2sql
async-paho-mqtt-client
hbmqtt
mqttools
backports.pbkdf2
gmqtt
async-timeout
WMI
pyroute2
pyuac
passlib
APScheduler
pyinstaller
tk
pycryptodomex
pycryptodome
netifaces
pyserial
SQLAlchemy-Utils
sqlalchemy-views